Conflicts:
	AndroidManifest.xml
	ChangeLog
	jni/Application.mk
	proguard.cfg
	src/org/geometerplus/fbreader/fbreader/FBReaderApp.java

+++ src/org/geometerplus/fbreader/fbreader/FBReaderApp.java
@@ -56,7 +56,7 @@ public final class FBReaderApp extends ZLApplication {
 		new ZLEnumOption<WordTappingAction>("Options", "WordTappingAction", WordTappingAction.startSelecting);
 
 	public final ZLColorOption ImageViewBackgroundOption =
-		new ZLColorOption("Colors", "ImageViewBackground", new ZLColor(127, 127, 127));
+		new ZLColorOption("Colors", "ImageViewBackground", new ZLColor(255, 255, 255));
 	public static enum ImageTappingAction {
 		doNothing, selectImage, openImageView
 	}
@@ -101,7 +101,7 @@ public final class FBReaderApp extends ZLApplication {
 	public final FBView BookTextView;
 	public final FBView FootnoteView;
 
-	public BookModel Model;
+	public volatile BookModel Model;
 
 	private final String myArg0;
 
@@ -230,7 +230,7 @@ public final class FBReaderApp extends ZLApplication {
 		FootnoteView.clearCaches();
 	}
 
-	private void openBookInternal(Book book, Bookmark bookmark) {
+	synchronized void openBookInternal(Book book, Bookmark bookmark) {
 		if (book != null) {
 			onViewChanged();
 
@@ -402,8 +402,8 @@ public final class FBReaderApp extends ZLApplication {
 		}
 	}
 
-	private void updateInvisibleBookmarksList(Bookmark b) {
-		if (Model.Book != null && b != null) {
+	private synchronized void updateInvisibleBookmarksList(Bookmark b) {
+		if (Model != null && Model.Book != null && b != null) {
 			for (Bookmark bm : Bookmark.invisibleBookmarks(Model.Book)) {
 				if (b.equals(bm)) {
 					bm.delete();
